Replace services component with component that uses new UnitDataList component
