--- src/views/UnitView/components/Services/index.js
@@ -1,4 +1,0 @@
-import { injectIntl } from 'react-intl';
-import Services from './Services';
-
-export default injectIntl(Services);
