+++ src/views/UnitView/components/UnitsServicesList/index.js
@@ -1,0 +1,4 @@
+import { injectIntl } from 'react-intl';
+import UnitsServicesList from './UnitsServicesList';
+
+export default injectIntl(UnitsServicesList);
